app/requirements.txt
pandas<2.2.0
pycaret==3.3.2
scikit-learn==1.4.2
scipy>=1.6.1
joblib==1.3.2
pyserial==3.5
numpy==1.26.4
matplotlib==3.7.5